ofi_net: Update ofi_bsock_send/ofi_bsock_sendv calls
Update the signature of the functions ofi_bsock_send and
ofi_bsock_sendv to match their recv/recvv counterparts.

The functions now return 0 on success and a negative value on
failure. The number of bytes consumed by the bsock interface for
the given call (whether the data is buffered in the bsock buffer
or sent on the socket) is returned in *len.

+++ include/ofi_net.h
@@ -457,9 +457,9 @@ int ofi_bsock_flush_sync(struct ofi_bsock *bsock);
 /* For sends started asynchronously, the return value will be -EINPROGRESS_ASYNC,
  * and len will be set to the number of bytes that were queued.
  */
-ssize_t ofi_bsock_send(struct ofi_bsock *bsock, const void *buf, size_t *len);
-ssize_t ofi_bsock_sendv(struct ofi_bsock *bsock, const struct iovec *iov,
-			size_t cnt, size_t *len);
+int ofi_bsock_send(struct ofi_bsock *bsock, const void *buf, size_t *len);
+int ofi_bsock_sendv(struct ofi_bsock *bsock, const struct iovec *iov,
+		    size_t cnt, size_t *len);
 int ofi_bsock_recv(struct ofi_bsock *bsock, void *buf, size_t *len);
 int ofi_bsock_recvv(struct ofi_bsock *bsock, struct iovec *iov,
 		    size_t cnt, size_t *len);
